app/requirements.txt
fastapi
uvicorn
onnxruntime
torchvision
pillow
numpy
python-multipart
